fix(settings):  move settings file path and reader outside of middleware package
Those are not strictly speaking redux middleware specific code and are used by some other part of the process.

This fixes an undefined reference issue on trying to import readSettings from middleware index file.
