+++ src/main/data/middlewares/index.ts
@@ -2,8 +2,6 @@ import { pipelineMiddleware } from './pipeline'
 import { settingsMiddleware } from './settings'
 import { updateMiddleware } from './update'
 
-export { readSettings } from './settings'
-
 export const middlewares = [
     settingsMiddleware,
     pipelineMiddleware,
